ai_3/requirements.txt
flask
flask-cors
groq
python-dotenv
gunicorn